Add new script for docker on mac, update README with setup instructions, mount dataset directory in docker run scripts, and compile opencv with gt2k enabled.

+++ docker_run.sh
@@ -1,3 +1,10 @@
 #!/usr/bin/env bash
-docker build -t darvis:latest . && docker run --mount type=bind,source="$(pwd)",target=/darvis -it darvis:latest
+if [ "$#" -lt 1 ]; then
+    echo "Usage: ./docker_run.sh [path to dataset directory]"
+    exit
+fi
+
+datasetdir=$1
+
+docker build -t darvis:latest . && docker run --mount type=bind,source="$(pwd)",target=/darvis --volume datasetdir:/datasets/ -it darvis:latest
 docker system prune -f
